extract input module
