+++ src/texture.rs
@@ -22,8 +22,7 @@ impl Texture {
 			sample_count,
 			dimension: wgpu::TextureDimension::D2,
 			format: Self::DEPTH_FORMAT,
-			usage: wgpu::TextureUsages::RENDER_ATTACHMENT
-                | wgpu::TextureUsages::TEXTURE_BINDING,
+			usage: wgpu::TextureUsages::RENDER_ATTACHMENT | wgpu::TextureUsages::TEXTURE_BINDING,
 		};
 		let texture = device.create_texture(&desc);
